requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
xarray>=2023.1.0
zarr>=2.14.0
dask>=2023.1.0
h5netcdf>=1.2.0
cfgrib>=0.9.10
eccodes>=1.5.0
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0
matplotlib>=3.6.0
seaborn>=0.12.0
tqdm>=4.64.0
pyyaml>=6.0
wandb>=0.15.0
faiss-gpu>=1.7.4
scikit-learn>=1.2.0
netcdf4>=1.6.0